Batch gitosis-admin updates up to the end of the request.
Instead of updating the projects as part of the observer lifecycle. We batch up changes and do them in one single update as part of the action lifecycle.
This does mean the admin updates no longer form part of the ActiveRecord transactiom, but is much much much faster esp. when adding groups which used to
repeat the admin update for each user in the group, over and over again!

I suspect some people will dislike the Thread local approach, but I was unsure how else to tackle this in Rails :(