Avoid un-neccessary gitolite configuration updates
There is some logic in update_repositories to avoid 'recursion'
however this does not protect against the case where an
observed change leads to multiple calls to update_repositories
for the same project (or projects.)

An example change that could cause this is is to add a *group*
with commit rights as a project's member.  The update_repositories
method will then get called for both the group's addition and
each subsequent member of that group.  This change avoids those
un-neccessary secondary calls.