Cache calls to 'entries'
The Hosting adapter causes multiple calls to be sent to 'entries' on
the git_adapter(repositories_controller_path#show_with_git_instructions
adds a second (non vanilla chiliproject) call.

As this method is very expensive, this has a performance impact we can
avoid!

It might be better to provide a dedicated method to the patched
git_adapter to determine whether a repository is 'empty' or not
rather than having to fetch the entries, but this will work for now.